refactor(core): consolidate the ci_* command modules under commands/ci
`ci_gate`, `ci_plan` and `ci_scope` each have exactly one consumer in the
workspace — `homeboy-cli`'s `commands/ci.rs` — and nothing inside
homeboy-core references any of them. They are the CI command's
implementation, not core infrastructure.

`ci_profile` is the one member of the `ci_*` group that genuinely belongs
in core: 12 files across homeboy-cli, homeboy-extension and homeboy-review
consume it. It stays.

Moved with `git mv` into `commands/ci/` and dropped the now-redundant
`ci_` prefix (`gate`, `plan`, `scope`). `ci_failure_log_triage` moves the
last hop from `commands/` into the same directory as `failure_log_triage`,
so the CI command is one subtree instead of four crate-root modules plus a
stray sibling. Each module's `#[cfg(test)] mod tests` moves with it.

Baseline rows repointed to the new paths for the same reason as the previous
commit: `validate_fingerprint_paths` rejects rows for files that no longer
exist. Count unchanged at 1137.

Refs #11143
