+++ crates/homeboy-cli/src/commands/mod.rs
@@ -230,7 +230,6 @@ pub mod bench;
 pub(crate) mod bench_summary;
 pub mod build;
 pub mod ci;
-pub(crate) mod ci_failure_log_triage;
 pub mod cleanup;
 pub mod component;
 pub mod config;
